feat: identity and authentication
This PR adds a basic authentication mechanism using random session tokens and
using Sign-in with Ethereum for users to authenticate with their wallets (including
Privy embedded wallets).

It also adds an identity mechanism by generating encryption and signing keys for each user,
which are stored on the sync server using a secret based on a wallet signature, similarly to XMTP v2.